+++ Tests/Add-GitItem.Tests.ps1
@@ -10,7 +10,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-& (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-LibGit2Test.ps1' -Resolve)
+& (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-GitAutomationTest.ps1' -Resolve)
 
 function Assert-FileNotStaged
 {
@@ -192,7 +192,7 @@ Describe 'Add-GitItem for an unmodified file' {
     $repoRoot = New-GitTestRepo
     Add-GitTestFile -Path 'file1' -RepoRoot $repoRoot
     Add-GitItem -Path 'file1','file1' -RepoRoot $repoRoot
-    Save-GitChange -Message 'Committing a file change' -RepoRoot $repoRoot
+    Save-GitCommit -Message 'Committing a file change' -RepoRoot $repoRoot
     Add-GitItem -Path 'file1' -RepoRoot $repoRoot
     Assert-FileNotStaged -Path 'file1' -RepoRoot $repoRoot
     It 'should not throw any errors' {