+++ Tests/Remove-GitItem.Tests.ps1
@@ -10,7 +10,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-& (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-LibGit2Test.ps1' -Resolve)
+& (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-GitAutomationTest.ps1' -Resolve)
 
 function GivenRepositoryWithFile
 {
@@ -24,7 +24,7 @@ function GivenRepositoryWithFile
         New-Item -Path (Join-Path -Path $repoRoot -ChildPath $item) -Type 'File'
     }
     Add-GitItem -RepoRoot $repoRoot -Path $Name
-    Save-GitChange -RepoRoot $repoRoot -Message 'Commit to add test files'
+    Save-GitCommit -RepoRoot $repoRoot -Message 'Commit to add test files'
 }
 
 function GivenIncorrectRepo