+++ Tests/Test-GitBranch.Tests.ps1
@@ -10,7 +10,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-& (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-LibGit2Test.ps1' -Resolve)
+& (Join-Path -Path $PSScriptRoot -ChildPath 'Initialize-GitAutomationTest.ps1' -Resolve)
 
 Describe 'Test-GitBranch when running from a valid git repository' {
     Clear-Error
@@ -18,7 +18,7 @@ Describe 'Test-GitBranch when running from a valid git repository' {
     $repo = New-GitTestRepo
     Add-GitTestFile -RepoRoot $repo -Path 'file1'
     Add-GitItem -Path (Join-Path -Path $repo -ChildPath 'file1') -RepoRoot $repo
-    Save-GitChange -RepoRoot $repo -Message 'file1 commit'
+    Save-GitCommit -RepoRoot $repo -Message 'file1 commit'
 
     It 'should return true if the branch name exists' {
         Test-GitBranch -RepoRoot $repo -Name 'master' | Should Be $true