+++ RELEASE_NOTES.md
@@ -62,7 +62,7 @@
 
  * Created `Add-GitItem` function for promoting new, untracked, and modified files/directories to the Git staging area so they can be committed. Implements the `git add` command.
  * Created `Get-GitRepository` function for getting an object representing a repository.
- * Created `Save-GitChange` function for commiting changes to a repository. Implements the `git commit` command.
+ * Created `Save-GitCommit` function for commiting changes to a repository. Implements the `git commit` command.
  * Created `Find-GitRepository` function for searching a directory and its parents (i.e. up it tree) for a Git repository.
  * Created `New-GitRepository` function for creating new Git repositories.
  * Created `Get-GitRepositoryStatus` function for getting the state of a repository's working directory and any items staged for the next commit.